requirements.txt
# Core dependencies
python-dotenv==1.0.0
pandas==2.0.3
requests==2.31.0
PyMuPDF==1.22.5  # For PDF processing
Pillow==10.0.0   # For image processing

# Google Gemini API
google-generativeai==0.3.1

# Azure OpenAI (for compatibility with existing scripts)
openai==1.3.5
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0

# Utilities
tqdm==4.66.1     # For progress bars
matplotlib==3.7.2  # For visualizations 